// features/products/constants/toppings.enum.ts
export enum Topping {
  Anchovies = 'anchovies',
  Bacon = 'bacon',
  Basil = 'basil',
  Chili = 'chili',
  Mozzarella = 'mozzarella',
  Mushrooms = 'mushrooms',
  Olive = 'olive',
  Onion = 'onion',
  Pepper = 'pepper',
  Pepperoni = 'pepperoni',
  Peppers = 'peppers',
  Sweetcorn = 'sweetcorn',
  Tomato = 'tomato',
  Tuna = 'tuna',
  Pineapple = 'pineapple',
  Ham = 'ham',
  Chicken = 'chicken',
  Beef = 'beef',
  Pork = 'pork',
}

// https://stackoverflow.com/a/47755096/426840